Stop linking OptiTrust library into scripts by using a plugin runner
Before this commit a lot of time is spent linking transformation scripts
because it statically include all the OptiTrust dependencies.
Since linking time is dominating in an iterative script edition session,
we need a way to reduce it.
OCaml plugins do not have to statically link all their libraries,
instead they use their host to grab them.
In this commit we create a simple plugin runner statically linked with
OptiTrust library and its dependencies, and we allow to build
transformation scripts as plugins.

TODO: Clean further the src/cparser directory since runner will embed
all of it in its binary.