Fix processing of Trm_let_mult in Trm.trm_map_vars
The original implementation was considering the different declarations in a multiple variable declaration in parallel. This was preventing us from doing for example 'int a = 1, b = a;' as all the initialization terms were processed at the beginning of the process, so 'a' in 'b = a' was not locally defined yet. To prevent this, in the new implementation, we interleave the processing of initialization terms with the binding of the variables being declared.