Take care of calls to functions we do not know the definition of
When in a function definition we make a call to a function we do not know the definition of, we have to consider all of its arguments as in-out dependencies. Therefore, during the constification process, the arguments or the aliases to the arguments of the currently processed function definition involved in such a function call must be unconstified.