Remove inconsistent Var_(im)mutable annotations on the AST
Transformations now consider that mutability can only come from indirection in the internal representation.
To know if a variable is mutable in the original source code, check if it is initialized from the ref operator.

This breaks apac/unfold_let_mult_const but this transformation is not removing ref operations as it probably should.