Remove efracs from resource sets
The current implementation created unsound usage maps.
Instead we chose an arbitrarily fixed fraction whenever we need to split a subfraction.
If we need to specify the subfractions to use we can always call the [ro_splitn] ghosts to split them before usage.

The transformation [Loop.fission_basic] will now fail if arbitrarily chosen variables are present at the split point.
An external transformation should be added later to specialize arbitrarily chosen fractions and restore the old
behaviour of combi versions.
Handling loop fission with arbitrarily chosen variables at the split point seems useful but is currently not used in the tests anyway.