Divide OptiTrust into 3 libraries (utils, ast and framework)
The C parser only depends on utils and ast and therefore should
not be recompiled when something in framework changes.
The C parser is now the only part of the project that depends on ClangML

Hopefully this should improve build times while working on the framework
and it removes the need for a manual `make clean_cache` when working
on the C parser.