Add an optional result value to sequences
A sequence may now be an expression with a result value.
The result of a sequence is expressed as an optional variable that
must be bound at toplevel of the sequence.
Return statements at the end of a function body are decoded as a result
value of the body sequence.
We use the GNU C "Statement Exprs" syntax extension to directly parse
sequences with result value.

Function.inline is refactored in the following way:
- Function.bind_args now perform the binding of arguments even when
checking resources.
- Function_basic.inline performs the raw inlining producing a nested
sequence expression to replace the call.
- Sequence.elim is able to eliminate the generated sequence in the
surrounding context.
- Old Function.bind and Function.bind_intro are completely removed and
replaced by equivalent Variable.bind.

For now Function.inline produces a slightly more ugly result in presence
of early return. This should be patched later when a proper support
for early return will be developped.