Fix bug leading us to put jumps into tasks
We were adding the `Singleton' attribute to the `IsJump' task candidates only when merging them, i.e. only when there were at least two of them at the end of a sequence. Now, we ensure the `Singleton' attribute even if there is only one `IsJump' task candidate at the end of a sequence.