Delete steps from trace as soon as possible and simplify Flags
The memory footprint of executions with Flags.save_steps = Steps_script
is now only propotional to the number of lines in the transformation
script.

Setting the execution mode does not modify other flags anymore:
instead some flags get accessors that depend on the execution mode.

This commit also simplifies the trace for backtracking steps
removing the intermediate group that was useless and the associated
invariants and exceptions.

Also add error message in trace view when the AST is missing