Check types during resource computation
This commit adds typechecking of program expressions and of resource formulas.
It also changes the encoding of allocations to use operators.

These changes forced to fix a lot of tests that did not typecheck.

Currently, structures are not handled and will now block the typechecker.
Also, there is much more work done at each resource computation and the
quadratic complexity of the typechecker is even more visible than before.