Remove ghost pure resources created by typing function arguments
This should reduce the size of contexts to what it was before adding
all program variables in context, by removing bindings for results of
unnamed subexpressions.

Some missed arithmetic simplifications are restored because of this.

Performance wise, this filtering should be improved when we have a
better data structure for storing contexts.