Restore struct handling
Construction of a struct or array is now done
through regular operators to avoid redundancy in
the code.

Struct accesses must now carry their type,
except in formulas where they must have type auto.

Parsing of initializer list is now semantic:
fields are reordered and zero-initializers are
inserted. This greatly simplifies the internal AST.
In the future, we could encode that as
annotations asking for a syntactic parsing and
doing the semantic translation in clang_to_ast.

We still have a var_id problem in micropic
Moreover, the unfolding lemmas should be generated
during typing instead of being admitted.